+++ src/main/java/com/skthvl/cinemetrics/repository/RatingRepository.java
@@ -3,13 +3,14 @@ package com.skthvl.cinemetrics.repository;
 import com.skthvl.cinemetrics.entity.Rating;
 import com.skthvl.cinemetrics.model.dto.RatingDto;
 import com.skthvl.cinemetrics.model.dto.TopRatedMovieDto;
+import java.math.BigInteger;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RatingRepository extends JpaRepository<Rating, Long> {
+public interface RatingRepository extends JpaRepository<Rating, BigInteger> {
 
   @Query(
       """
